Update BVB_WebConfig_OTA_V7.ino
The includes are case sensitive on Linux and OS X systems, so fixed that. 
The other thing is that the <credentials.h> file is not available in the repository. I am guessing that they ought to contain default values for ASP_SSID and ASP_password. So getting rid of this one means a later compile error on ASP_SSID and ASP_password. Fixed that by defaulting to defaults if they are not defined.